server/requirements.txt
fastapi
uvicorn
watchdog
python-socketio
typer
ruff